refactor: MailService에 Facade 계층 도입 #144
